chore: refactor logic out of router and into controllers
